More support for terabit beta Added serverids to help find out info with mutiple instances

+++ entrypoint.sh
@@ -1,5 +1,10 @@
-#!/bin/sh
-pip3 install -r ./requirements.txt
+#!/bin/bash
+if ! test -f "./tuberepair/serverID.txt"; then
+    pip3 install -r ./requirements.txt
+    useradd tubeuser
+    chown -R tubeuser:tubeuser ./tuberepair
+fi
+su tubeuser
 cd ./tuberepair
 echo 'Starting TubeRepair'
 python3 ./main.py
